+++ CHANGELOG.md
@@ -4,14 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.5.0] - 2019-05-20
+### Added
+- Body parsing support for `application/x-www-form-urlencoded`
+- `event.rawBody` as original, unparsed body when `event.body` is parsed
+### Fixed
+- Check `Content-Type` for `application/json` before parsing as JSON
+
 ## [2.4.1] - 2019-05-01
 ### Fixed
-Documentation errors
+- Documentation errors
 
 ## [2.4.0] - 2018-12-01
 ### Fixed
-Documentation errors
+- Documentation errors
 
 ## [2.3.0] - 2018-10-8
 ### Added
-Initial OSS release
+- Initial OSS release